# Conflicts:
#	DataPoolHandler.py
#	Utility.py

+++ DataPoolHandler.py
@@ -1,21 +1,18 @@
 #==================================================================
 #==================================================================
-from genericpath import isdir, isfile
+
 import os
-from posixpath import basename
-from typing import Dict
 import cv2
 import pickle
 from PyQt5.QtWidgets import QMessageBox
-from PyQt5.QtCore import QObject, pyqtSlot, pyqtSignal
+from PyQt5.QtCore import QObject,  pyqtSignal
 import numpy as np
 from shutil import copyfile
-from glob import glob
 import ptvsd
+#import ptvsd
 #import ptvsd
 import pydicom
 from pydicom import dcmread
-import Class
 from Utility import *
 import Config
 from Strategy import get_grad_embeddings, get_cluster_centers
@@ -178,21 +175,22 @@ class DataPoolHandler(QObject):
         if os.path.exists(os.path.join(Config.PROJECT_ROOT, 'labels', f'{file_name}.meta')):
             meta_file = pickle.load(open(os.path.join(Config.PROJECT_ROOT, 'labels', f'{file_name}.meta'), 'rb'));
             for k in meta_file.keys():
-                if k != 'rot' and k!= 'exp':
+                if k != 'misc' and k!='rot' and k!='exp':
                     os.remove(os.path.join(Config.PROJECT_ROOT, 'labels',meta_file[k][2]));
         
             os.remove(os.path.join(Config.PROJECT_ROOT, 'labels', f'{file_name}.meta'));
         
         self.save_project_signal.emit(False);
 
-    def submit_label(self,  arr, rot, exp):
+    def submit_label(self,  arr, misc):
+        #ptvsd.debug_this_thread();
         self.__data_list[self.__current_radiograph][0] = "labeled";
 
         path_tmp = self.__current_radiograph.replace('\\','/');
         #save label to labels folder and save meta data about radiograph
         file_name = os.path.basename(path_tmp);
         file_name = file_name[0:file_name.find('.')];
-        data_dict = dict({'rot' : rot, 'exp': exp});
+        data_dict = dict({'misc' : misc});
 
         for l in range(len(arr)):
             layer = arr[l][0];
@@ -230,12 +228,15 @@ class DataPoolHandler(QObject):
         #rename image, if image with new name already exists, skip renaming
         if os.path.exists(f'{Config.PROJECT_ROOT}\\images\\{new_name}{ext}') is False:
             os.rename(f'{Config.PROJECT_ROOT}\\images\\{orig_name}', f'{Config.PROJECT_ROOT}\\images\\{new_name}{ext}');
+        #remove original image
+        if os.path.exists(f'{Config.PROJECT_ROOT}\\images\\{orig_name}') is True:
+            os.remove(f'{Config.PROJECT_ROOT}\\images\\{orig_name}');
 
         if self.__data_list[f"{new_name}{ext}"][0] == 'labeled':
             #rename all_labels
             meta_file = pickle.load(open(os.path.join(Config.PROJECT_ROOT, 'labels', f'{orig_name_we}.meta'), 'rb'));
             for m in meta_file.keys():
-                if m != 'rot' and m != 'exp':
+                if m != 'misc' and m!= 'rot' and m!='exp':
                     mask_name = meta_file[m][2];
                     mask_idx = mask_name[mask_name.find('_'):];
                     new_mask_name = f"{new_name}{mask_idx}";
